baseline/source_simplification/requirements.txt
# Core dependencies
langchain-core>=0.2.32
langchain>=0.2.14
langchain_huggingface>=0.0.3
langchain_community
python-dotenv>=1.0.1
ragas==0.1.21
PyYAML>=6.0.2
tqdm>=4.66.4
hydra-core>=1.3.2
datasets>=3.5.0
mlflow>=2.17.2
beautifulsoup4>=4.13.3
omegaconf>=2.3.0
pandas>=2.0.0
pydantic>=2.0.0
openai
rank_bm25
faiss-cpu
sentence-transformers

# Runtime dependencies
torch
transformers
accelerate
uvicorn
